quickboosters/api/chat/requirements.txt
flask
flask-socketio
flask-sqlalchemy
python-dotenv